+++ setup.cfg
@@ -79,14 +79,13 @@ install_requires =
     pooch
     pyproj>=3.3
     pyvista==0.37.0
-    pyvistaqt
     vtk
 package_dir =
     =src
 packages = find_namespace:
 python_requires =
     >=3.8
-zip_safe = False 
+zip_safe = False
 
 [options.entry_points]
 console_scripts =
